+++ R/flash_flights_parse.R
@@ -20,11 +20,11 @@
 #'   \code{add_row_numbers}
 #' @return returns a data frame with split times and row numbers
 #'
-#' @seealso \code{flights_parse_flash} runs inside \code{\link{flash_parse}} on
+#' @seealso \code{rounds_parse_flash} runs inside \code{\link{flash_parse}} on
 #'   the output of \code{\link{read_results}} with row numbers from
 #'   \code{\link{add_row_numbers}}
 
-flash_flights_parse <- function(text) {
+flash_rounds_parse <- function(text) {
   #### Testing ####
   # file <- "http://leonetiming.com/2019/Indoor/GregPageRelays/Results.htm"
   # file <-
@@ -55,15 +55,15 @@ flash_flights_parse <- function(text) {
     .[purrr::map_lgl(., ~ !any(stringr::str_detect(., "(?<=[\U2013]) +X ")))] %>%  # remove special case where in vertical jumps someone passes a round then fails leaving "XX-  X" and that X gets picked up like it would for a horizontal event
     .[purrr::map_lgl(., ~ !any(stringr::str_detect(., "(?<=O) +X ")))] # remove special case where in vertical jumps someone passes a round then makes one attempt and passes leaving "O  X" and that X gets picked up like it would for a horizontal event
 
-  #### collect row numbers from rows containing flights ####
+  #### collect row numbers from rows containing rounds ####
   row_numbs <- text %>%
     .[purrr::map_lgl(., stringr::str_detect, attempt_string_flash)] %>%
     str_extract("\\d{1,}$")
 
-  #### pull out rows containing flights ####
+  #### pull out rows containing rounds ####
 
   suppressWarnings(
-    data_flights <- text %>%
+    data_rounds <- text %>%
       .[purrr::map_lgl(., stringr::str_detect, attempt_string_flash)] %>%
       # .[purrr::map_lgl(., ~ stringr::str_detect(., "^\\d", negate = TRUE))] %>% # removes rows that start with a place, to remove main results and scores with decimal places (5 Alfred U. 2.50 etc.)
       stringr::str_extract_all(attempt_string_flash, simplify = TRUE) %>%
@@ -71,39 +71,39 @@ flash_flights_parse <- function(text) {
   )
 
   #### reattach row numbers ####
-  data_flights <- cbind(row_numbs, data_flights) %>%
+  data_rounds <- cbind(row_numbs, data_rounds) %>%
     as.data.frame() %>%
     dplyr::na_if("") %>%
     dplyr::rename(V1 = row_numbs) # for list_sort, needs V1 to be row numbers, but named V1
 
   if(any(stringr::str_detect(text, "Scored")) == TRUE){ # gets rid of team scores in vertical jump events
     row_score <- min(as.numeric(stringr::str_extract(text[stringr::str_detect(text, "Scored")], "\\d{1,}$")))
-    data_flights <- data_flights %>%
+    data_rounds <- data_rounds %>%
       dplyr::filter(as.numeric(row_numbs) < row_score)
   }
 
   #### reattach row numbers ####
 
-  data_flights <- data_flights %>%
+  data_rounds <- data_rounds %>%
     lines_sort(min_row = min(as.numeric(row_numbs))) %>%
     dplyr::mutate(Row_Numb = as.numeric(Row_Numb))
 
   #### rename columns V1, V2 etc. at Attempt_1, Attempt_2 etc. ####
   old_names <-
-    names(data_flights)[grep("^V", names(data_flights))]
+    names(data_rounds)[grep("^V", names(data_rounds))]
   new_names <-
-    paste("Flight", seq(1, length(names(data_flights)) - 1), sep = "_")
+    paste("Round", seq(1, length(names(data_rounds)) - 1), sep = "_")
 
-  data_flights <- data_flights %>%
+  data_rounds <- data_rounds %>%
     dplyr::rename_at(dplyr::vars(dplyr::all_of(old_names)), ~ new_names)
 
   if (sum(suppressWarnings(str_detect(text, "\\d\\.\\d{2}m"))) >= 1) {
-    # keeps running times like 10.34 from getting into flights
-    return(data_flights)
+    # keeps running times like 10.34 from getting into rounds
+    return(data_rounds)
   } else {
-    data_flights <- data.frame(Row_Numb = character(),
+    data_rounds <- data.frame(Row_Numb = character(),
                                 stringsAsFactors = FALSE)
-    return(data_flights)
+    return(data_rounds)
   }
 
 }